refactor: Cleanly separate between transport and domains.
To discuss around clean archi. I rework code archi to put
transport in a transport module and domains in a domain module.

As far as I go with it, I like to separate between:
- Domains: all objects used anywhere in the app
- Transport: all features specific to how I communicate with
            the outside world
- Usecase: what I do
- Modules: how do I do what I should
