+++ .env.sample
@@ -1,1 +1,2 @@
 DISCORD_BOT_SECRET_KEY=YOUR_DISCORD_SECRET_KEY
+DISCORD_BOT_ADMIN_ID=SOME_DISCORD_USER_ID
