+++ app/domains/exceptions.py
