+++ app/transport/__init__.py
@@ -1,0 +1,1 @@
+from .discord_client import DiscordClient as Discord
