+++ app/usecases/__init__.py
@@ -1,1 +1,8 @@
-from .miscellaneous import Miscellaneous # expose Miscellaneous in module
+"""
+Usecases regroup code describing the Domain part of the app.
+It has no tech related information and should juste contain
+code to describe clearly what has to be done for each known
+actions.
+"""
+
+from .miscellaneous import Miscellaneous  # expose Miscellaneous in module
