+++ pytest.ini
@@ -8,4 +8,4 @@ filterwarnings =
 [pylama]
 format = pylint
 linters = pylint,mccabe
-ignore = F0401, C0111, E731
+ignore = F0401, C0111, E731, C0114
